9 module endpoints (#17)

+++ FeedTrac.Server/Controllers/ModuleController.cs
@@ -1,7 +1,10 @@
 ﻿using FeedTrac.Server.Database;
+using FeedTrac.Server.Models.Responses.Modules;
 using FeedTrac.Server.Services;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace FeedTrac.Server.Controllers;
 
@@ -15,12 +18,14 @@ public class ModuleController : Controller
 
     private readonly ApplicationDbContext _context;
     private readonly UserService _userService;
+    private readonly UserManager<ApplicationUser> _userManager;
     private readonly ModuleService _moduleService;
 
-    public ModuleController(ApplicationDbContext context, UserService userService, ModuleService moduleService)
+    public ModuleController(ApplicationDbContext context, UserService userService, UserManager<ApplicationUser> userManager, ModuleService moduleService)
     {
         _context = context;
         _userService = userService;
+        _userManager = userManager;
         _moduleService = moduleService;
     }
 
@@ -28,19 +33,17 @@ public class ModuleController : Controller
     /// Retrieves a list of modules that the user is signed up to
     /// </summary>
     /// <example> GET /modules </example>
-    /// <returns>
-    ///     <b>200:</b> Returns the list of modules <br/>
-    ///     <b>401:</b> The user is not signed in <br/>
-    /// </returns>
+    /// <response code="200">Returns user's modules</response>
+    /// <response code="404">The user is not signed in</response>
     [HttpGet]
-    [ProducesResponseType(typeof(List<Module>), 200)]
+    [ProducesResponseType(typeof(UserModulesResponse), 200)]
     [ProducesResponseType(401)]
-    public async Task<IActionResult> GetModules()
+    public async Task<IActionResult> GetUserModules()
     {
         try
         {
             var modules = await _moduleService.GetUserModulesAsync();
-            return Ok(modules);
+            return Ok(new UserModulesResponse(modules));
         }
         catch (Exception e)
         {
@@ -49,17 +52,33 @@ public class ModuleController : Controller
     }
 
     /// <summary>
-    /// API endpoint for joining the current user to a module
+    /// Retrieves every module
+    /// </summary>
+    /// <example> GET /modules </example>
+    /// <response code="200">Returns user's modules</response>
+    /// <response code="404">The user does not have permissions</response>
+    [HttpGet("all")]
+    [Authorize(Roles = "Admin")]
+    [ProducesResponseType(typeof(UserModulesResponse), 200)]
+    [ProducesResponseType(404)]
+    public async Task<IActionResult> GetAllModules()
+    {
+        var modules = await _context.Modules.ToListAsync();
+        return Ok(new UserModulesResponse(modules));
+    }
+
+    /// <summary>
+    /// Joins signed in user to a module
     /// </summary>
     /// <param name="joinCode"></param>
     /// <example> POST modules/join?joinCode=xxxxxx </example>
-    /// <returns>
-    ///     <b>200:</b> The user has been successfully joined to the module <br/>
-    ///     <b>400:</b> Failed to join the module <br/>
-    ///     <b>401:</b> The user is not signed in <br/>
-    /// </returns>
-    [HttpPost]
-    [Route("join")]
+    /// <response code="200">The user has been added</response>
+    /// <response code="400">Failed to join the module</response>
+    /// <response code="404">The user is not signed in</response>
+    [HttpPost("join")]
+    [Authorize(Roles = "Student")]
+    [ProducesResponseType(typeof(ModuleResponse), 200)]
+    [ProducesResponseType(404)]
     public async Task<IActionResult> JoinModule(string joinCode)
     {
         if (_userService.GetCurrentUserId() == null)
@@ -68,7 +87,7 @@ public class ModuleController : Controller
         try
         {
             var module = await _moduleService.JoinModule(joinCode);
-            return Ok(module);
+            return Ok(new ModuleResponse(module));
         }
         catch(Exception e)
         {
@@ -77,21 +96,85 @@ public class ModuleController : Controller
     }
 
     /// <summary>
-    /// Get info about a module
+    /// Assigns a Teacher to a module
+    /// </summary>
+    /// <param name="moduleId"></param>
+    /// <param name="teacherId"></param>
+    /// <response code="200">The teacher has been assigned</response>
+    /// <response code="404">The module or teacher was not found</response>
+    [HttpPost("/{moduleId}/assignTeacher/byTeacherId")]
+    [Authorize(Roles = "Admin")]
+    [ProducesResponseType(typeof(string), 200)]
+    [ProducesResponseType(404)]
+    public async Task<IActionResult> AssignTeacherById(string moduleId, string teacherId)
+    {
+        var module = await _context.Modules.FirstOrDefaultAsync(m => m.Id == int.Parse(moduleId));
+        if (module == null)
+            return NotFound("Module not found");
+
+        var user = await _context.Users.FirstOrDefaultAsync(u => u.Id == teacherId);
+        if (user == null)
+            return NotFound("User not found");
+
+        module.TeacherModule.Add(new TeacherModule
+        {
+            UserId = user.Id,
+            ModuleId = module.Id
+        });
+
+        await _context.SaveChangesAsync();
+        return Ok("Teacher assigned");
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="moduleId"></param>
+    /// <param name="teacherEmail"></param>
+    /// <returns></returns>
+    [HttpPost("/{moduleId}/assignTeacher/byTeacherEmail")]
+    [Authorize(Roles = "Admin")]
+    [ProducesResponseType(typeof(string), 200)]
+    [ProducesResponseType(404)]
+    public async Task<IActionResult> AssignTeacherByEmail(string moduleId, string teacherEmail)
+    {
+        var module = await _context.Modules.FirstOrDefaultAsync(m => m.Id == int.Parse(moduleId));
+        if (module == null)
+            return NotFound("Module not found");
+
+        var user = await _context.Users.FirstOrDefaultAsync(u => u.Email == teacherEmail);
+        if (user == null)
+            return NotFound("Teacher not found");
+        List<string> userRoles = (await _userManager.GetRolesAsync(user)).ToList();
+        if ((userRoles.Contains("Teacher") || userRoles.Contains("Admin")) == false)
+            return BadRequest("User is not a teacher");
+
+        module.TeacherModule.Add(new TeacherModule
+        {
+            UserId = user.Id,
+            ModuleId = module.Id
+        });
+
+        await _context.SaveChangesAsync();
+        return Ok("Teacher assigned");
+    }
+
+
+    /// <summary>
+    /// Get module info
     /// </summary>
     /// <param name="id">The id of the module</param>
-    /// <example> GET modules/[moduleId] </example>
-    /// <returns>
-    ///     <b>200:</b> Returns the module <br/>
-    ///     <b>404:</b> The module was not found <br/>
-    /// </returns>
+    /// <response code="200">Returns the module</response>
+    /// <response code="404">The module was not found</response>
     [HttpGet("{id}")]
+    [ProducesResponseType(typeof(ModuleResponse), 200)]
+    [ProducesResponseType(404)]
     public async Task<IActionResult> GetModule(int id)
     {
         try
         {
             var module = await _moduleService.GetModuleAsync(id);
-            return Ok(module);
+            return Ok(new ModuleResponse(module));
         }
         catch
         {
@@ -103,13 +186,18 @@ public class ModuleController : Controller
     /// Delete a module
     /// </summary>
     /// <param name="id">The id of the module</param>
-    /// <returns>
-    ///     <b>200:</b> The module has been deleted <br/>
-    ///     <b>400:</b> Failed to delete the module <br/>
-    /// </returns>
+    /// <response code="200">The module was deleted</response>
+    /// <response code="404">The module was not found or the user does not have permission</response>
     [HttpDelete("{id}")]
+    [Authorize(Roles = "Admin")]
+    [ProducesResponseType(200)]
+    [ProducesResponseType(404)]
     public async Task<IActionResult> DeleteModule(int id)
     {
+        var module = await _context.Modules.FirstOrDefaultAsync(m => m.Id == id);
+        if (module == null)
+            return NotFound("Module not found");
+
         try
         {
             await _moduleService.DeleteModuleAsync(id);
@@ -125,16 +213,16 @@ public class ModuleController : Controller
     /// Create a module
     /// </summary>
     /// <param name="name"></param>
-    /// <returns></returns>
-    [HttpPost]
-    [Route("create")]
+    /// <response code="200">The module was created</response>
+    /// <response code="404">The user does not have permission</response>
+    [HttpPost("/create")]
+    [Authorize(Roles = "Admin")]
+    [ProducesResponseType(typeof(ModuleResponse), 200)]
+    [ProducesResponseType(404)]
     public async Task<IActionResult> CreateModule(string name)
     {
         if (_userService.GetCurrentUserId() == null)
             return Unauthorized("You must be signed in to create a module");
-        // TODO: Implement lecturer role check
-        //if (_userService.GetCurrentUserAsync().Result.Role != 0)
-        //    return Unauthorized("You must be a lecturer to create a module");
 
         var module = await _moduleService.CreateModuleAsync(name);
         return Ok(module);
